added webpack & modules
