+++ public/index.html
@@ -3,7 +3,8 @@
 
 <head>
   <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0" />
+  <meta name="viewport"
+    content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0" />
   <meta http-equiv="X-UA-Compatible" content="ie=edge" />
   <link rel="stylesheet" href="css/normalize.css" />
   <link rel="stylesheet" href="css/style.css" />
@@ -15,7 +16,8 @@
     <svg>
       <symbol id="wave" viewBox="0 0 159 10" preserveAspectRatio="none">
         <g fill="none" fill-rule="nonzero" stroke-linecap="square" stroke-width="5">
-          <path d="M4 4l9.257 2.463L21.367 4l7.927 2.463L38.736 4l9.117 2.463L56.103 4l8.685 2.463L73.472 4l8.684 2.463L90.84 4l8.684 2.463L108.208 4l8.684 2.463L125.576 4l8.684 2.463L142.943 4M146.423 4l9.257 2.463" />
+          <path
+            d="M4 4l9.257 2.463L21.367 4l7.927 2.463L38.736 4l9.117 2.463L56.103 4l8.685 2.463L73.472 4l8.684 2.463L90.84 4l8.684 2.463L108.208 4l8.684 2.463L125.576 4l8.684 2.463L142.943 4M146.423 4l9.257 2.463" />
         </g>
       </symbol>
     </svg>
@@ -37,7 +39,8 @@
     </section>
 
     <section class="main__search search container">
-      <input type="text" id="search__input" class="search__input search__input--hidden" placeholder="START TYPING — SEARCH BY WORD, #HASHTAG OR DATE" />
+      <input type="text" id="search__input" class="search__input search__input--hidden"
+        placeholder="START TYPING — SEARCH BY WORD, #HASHTAG OR DATE" />
       <label class="visually-hidden" for="search__input">Поиск</label>
     </section>
 
@@ -98,7 +101,7 @@
       <button class="load-more" type="button">load more</button>
     </section>
   </main>
-  <script src="../src/main.js"></script>
+  <script type="text/javascript" src="js/bundle.js"></script>
 </body>
 
 </html>
