+++ src/appendElement.js
@@ -1,0 +1,3 @@
+export default (node, element) => {
+  node.insertAdjacentHTML(`beforeend`, element);
+};
