+++ src/utils.js
@@ -1,0 +1,5 @@
+const getRandomNumber = (min, max) => {
+  return Math.random() * (max - min) + min;
+};
+
+export {getRandomNumber};
