+++ webpack.config.js
@@ -1,0 +1,29 @@
+const path = require(`path`);
+const HtmlWebpackPlugin = require(`html-webpack-plugin`);
+
+module.exports = {
+  mode: `development`,
+  entry: `./src/main.js`,
+  output: {
+    filename: `bundle.js`,
+    path: path.join(__dirname, `public/js`)
+  },
+  module: {
+    rules: [{
+      test: /\.js$/,
+      use: `babel-loader`
+    }]
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: `./public/index.html`
+    })],
+  devtool: `source-map`,
+  devServer: {
+    contentBase: path.join(__dirname, `public`),
+    publicPath: 'http: //localhost:8080/',
+    hot: true,
+    compress: true,
+    watchContentBase: true
+  }
+};
